Move AddCommentService from Utility to Services namespace
