+++ src/Rector/Deprecation/DrupalSetMessageRector.php
@@ -2,7 +2,7 @@
 
 namespace DrupalRector\Rector\Deprecation;
 
-use DrupalRector\Utility\AddCommentService;
+use DrupalRector\Services\AddCommentService;
 use DrupalRector\Utility\FindParentByTypeTrait;
 use PhpParser\Comment;
 use PhpParser\Node;
@@ -34,7 +34,7 @@ final class DrupalSetMessageRector extends AbstractRector
     use FindParentByTypeTrait;
 
     /**
-     * @var \DrupalRector\Utility\AddCommentService
+     * @var \DrupalRector\Services\AddCommentService
      */
     private AddCommentService $commentService;
